requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
blinker==1.6.2; python_version >= '3.7'
click==8.1.6; python_version >= '3.7'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask==2.3.2
flask-restful==0.3.10
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0
waitress==2.1.2
watchdog==3.0.0
werkzeug==2.3.6; python_version >= '3.8'
